requirements.txt
certifi==2018.1.18
chardet==3.0.4
enforce==0.3.4
idna==2.6
PyYAML==3.12
requests==2.18.4
urllib3==1.22
wrapt==1.10.11
